+++ tests/testthat/test-manage_lib.R
@@ -5,7 +5,7 @@ dir.create(tmp, showWarnings = F)
 
 test_that("check_lib() finds complete library", {
   skip_on_cran()
-  skip_if_offline(host = "osf.io")
+  skip_if_offline(host = "d2jrxerjcsjhs7.cloudfront.net")
   skip_if_not(testthat:::on_ci(), "Not on CI")
   
   get_lib(type = c("derivative"), path = tmp, aws = TRUE, revision = "DtlNCNhOGdtCJqRCN3F8jRs732t_yZLM")
@@ -14,10 +14,12 @@ test_that("check_lib() finds complete library", {
   get_lib(type = c("nobaseline"), path = tmp, aws = TRUE)
   check_lib(type = c("nobaseline"), path = tmp) |>
       expect_silent()
-  get_lib(type = c("medoid_derivative"), path = tmp)
+  get_lib(type = c("medoid_derivative"), path = tmp, aws = TRUE,
+          revision = "iThmNyMeUKhkWMvbBxQqpf1sESdQBFTs")
   check_lib(type = c("medoid_derivative"), path = tmp) |>
     expect_silent()
-  get_lib(type = c("medoid_nobaseline"), path = tmp, revision = 1)
+  get_lib(type = c("medoid_nobaseline"), path = tmp, aws = TRUE,
+          revision = "CLJCDpeFCMZw4hFUW4Y1QFT2cj23W1Yz")
   check_lib(type = c("medoid_nobaseline"), path = tmp) |>
       expect_silent()
   check_lib(type = "raw", path = tmp) |>
@@ -26,7 +28,7 @@ test_that("check_lib() finds complete library", {
 
 test_that("load_lib() works with complete library", {
   skip_on_cran()
-  skip_if_offline(host = "osf.io")
+  skip_if_offline(host = "d2jrxerjcsjhs7.cloudfront.net")
   skip_if_not(testthat:::on_ci(), "Not on CI")
 
   tl <- load_lib(type = "derivative", path = tmp) |>